+++ rkt/api_service.go
@@ -15,8 +15,10 @@
 package main
 
 import (
+	"bufio"
 	"encoding/json"
 	"fmt"
+	"io"
 	"net"
 	"os"
 	"os/signal"
@@ -24,12 +26,15 @@ import (
 	"path/filepath"
 	"strings"
 	"syscall"
+	"time"
 
 	"github.com/appc/spec/schema"
 	"github.com/appc/spec/schema/types"
+	"github.com/coreos/go-systemd/sdjournal"
 	"github.com/coreos/rkt/api/v1alpha"
 	"github.com/coreos/rkt/common"
 	"github.com/coreos/rkt/pkg/set"
+	stage1common "github.com/coreos/rkt/stage1/init/common"
 	"github.com/coreos/rkt/store"
 	"github.com/coreos/rkt/version"
 	"github.com/spf13/cobra"
@@ -655,8 +660,106 @@ func (s *v1AlphaAPIServer) InspectImage(ctx context.Context, request *v1alpha.In
 	return &v1alpha.InspectImageResponse{Image: image}, nil
 }
 
+// LogReadWriter is a wrapper around a GRPC streaming server.
+// Implements io.Writer interface for streaming logs back to
+// clients, and also implements the io.ReaderFrom interface
+// in order to efficiently copy log lines from the journal.
+type LogReadWriter struct {
+	server v1alpha.PublicAPI_GetLogsServer
+	lines  []string
+}
+
+// Write implements the io.Writer interface and is used
+// to stream log lines back to the client.
+func (rw LogReadWriter) Write(b []byte) (int, error) {
+	lines := strings.Split(string(b), "\n")
+	err := rw.server.Send(&v1alpha.GetLogsResponse{Lines: lines})
+	if err != nil {
+		return 0, err
+	}
+	return len(b), nil
+}
+
+// ReadFrom implements the io.ReaderFrom interface and is used
+// when passed into io.Copy to efficiently copy lines from a
+// journal file.
+func (rw *LogReadWriter) ReadFrom(r io.Reader) (int64, error) {
+	var n int64
+	rdr := bufio.NewReader(r)
+	for {
+		line, err := rdr.ReadString('\n')
+		if err != nil {
+			if err == io.EOF {
+				return n, nil
+			}
+			return n, err
+		}
+		n += int64(len(line))
+		rw.lines = append(rw.lines, line)
+	}
+}
+
 func (s *v1AlphaAPIServer) GetLogs(request *v1alpha.GetLogsRequest, server v1alpha.PublicAPI_GetLogsServer) error {
-	return fmt.Errorf("not implemented yet")
+	uuid, err := types.NewUUID(request.PodId)
+	if err != nil {
+		return err
+	}
+	pod, err := getPod(uuid)
+	if err != nil {
+		return err
+	}
+
+	stage1Path := "stage1/rootfs"
+	if pod.usesOverlay() {
+		stage1TreeStoreID, err := pod.getStage1TreeStoreID()
+		if err != nil {
+			return err
+		}
+		stage1Path = fmt.Sprintf("/overlay/%s/upper/", stage1TreeStoreID)
+	}
+
+	path := filepath.Join(getDataDir(), "/pods/run/", request.PodId, stage1Path, "/var/log/journal/")
+	jconf := sdjournal.JournalReaderConfig{
+		Path: path,
+	}
+	if request.AppName != "" {
+		acname, err := types.NewACName(request.AppName)
+		if err != nil {
+			return err
+		}
+		jconf.Matches = []sdjournal.Match{
+			{
+				Field: sdjournal.SD_JOURNAL_FIELD_SYSTEMD_UNIT,
+				Value: stage1common.ServiceUnitName(*acname),
+			},
+		}
+	}
+
+	if request.SinceTime != 0 {
+		t := time.Unix(request.SinceTime, 0)
+		jconf.Since = time.Since(t)
+	}
+
+	if request.Lines != 0 {
+		jconf.NumFromTail = uint64(request.Lines)
+	}
+
+	jr, err := sdjournal.NewJournalReader(jconf)
+	if err != nil {
+		return err
+	}
+	defer jr.Close()
+
+	w := LogReadWriter{server: server}
+	if request.Follow {
+		return jr.Follow(nil, w)
+	}
+
+	_, err = io.Copy(&w, jr)
+	if err != nil {
+		return err
+	}
+	return server.Send(&v1alpha.GetLogsResponse{Lines: w.lines})
 }
 
 func (s *v1AlphaAPIServer) ListenEvents(request *v1alpha.ListenEventsRequest, server v1alpha.PublicAPI_ListenEventsServer) error {